requirements.txt
Flask==3.0.3
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
paho-mqtt==2.1.0
eventlet==0.36.1